+++ AddOns/!KalielsTracker/Modules/QuestLog.lua
@@ -106,7 +106,7 @@ local function SetHooks()
 	end
 
 	hooksecurefunc("QuestMapQuestOptions_TrackQuest", function(questID)
-		if not C_SuperTrack.GetSuperTrackedQuestID() then
+		if db.questAutoFocusClosest and not C_SuperTrack.GetSuperTrackedQuestID() then
 			KT.QuestSuperTracking_ChooseClosestQuest()
 		end
 	end)